refactor: extracted context from hooks file.
